move rows to separate file
